Add some casts to avoid warnings on 64bit systems
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/branches/releases/1.6@2633 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CImageLoaderPNG.cpp
@@ -38,7 +38,7 @@ void PNGAPI user_read_data_fcn(png_structp png_ptr, png_bytep data, png_size_t l
 
 	// changed by zola {
 	io::IReadFile* file=(io::IReadFile*)png_ptr->io_ptr;
-	check=(png_size_t) file->read((void*)data,length);
+	check=(png_size_t) file->read((void*)data,(u32)length);
 	// }
 
 	if (check != length)
